simplify svn module
